+++ CHANGES.md
@@ -30,6 +30,7 @@
 -   Refactor 'inputTypes', delete 'getInputType' #138
 -   Refactor tableView structure to fit pagination #135
 -   Add search component #147
+-   Optimize select input type #146
 
 ## Version 1.4.0